Extract Package component to make package reusable for both packages and wishlist
