+++ src/index.css
@@ -124,6 +124,7 @@ main button {
     justify-content: flex-start;
     padding: 0;
   }
+  
   main button {
     padding: 0;
     width: 100%;
